Add swsusp image restore command
This patch adds a command to load and boot a swsusp image. It requires an arch specific
call_with_stack function as well as some small kernel changes.

It requires the kernel to add a cpu_resume function to the swsusp_info block and also requies
that the kernel restore it's nosave pages.

This patch supports both uncompressed and LZO compressed hibernation images.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>